requirements.txt
httpx>=0.27.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pyyaml>=6.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
jinja2>=3.1.0
schedule>=1.2.0
tenacity>=8.0.0
